refactor(agent-v4): 收起坞的空当算式搬进自己的文件
上一条修复把 `transportClearanceFrom` 导出在 `AgentPanelV4Dock.tsx` 上，那个文件此前
只导出组件——多一个函数导出会让 Fast Refresh 失效，`lint:ci` 的 81 条警告棘轮当场红
（82 > 81）。纯换算本来就该单独一个文件才好被单测钉住，顺手归位，不改行为。
